locking/seqlock: Split out seqlock_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

Trimming down sched.h dependencies: we don't want to include more than
the base types.

Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Will Deacon <will@kernel.org>
Cc: Waiman Long <longman@redhat.com>
Cc: Boqun Feng <boqun.feng@gmail.com>
(cherry picked from commit f038cc1379c0ff462d83895cae8beb75a0f6bf02)
